src/retrieval/requirements.txt
torch
torchvision
transformers
accelerate
fastapi
uvicorn[standard]
pillow
rasterio
mercantile
numpy
qdrant-client
pydantic
pyyaml
tqdm
pandas
requests
sentencepiece
pyarrow
psycopg[binary]
sentence-transformers
